+++ extra/src/main/scala/ai/lum/odinson/extra/utils/ProcessorsUtils.scala
@@ -1,22 +1,20 @@
-package ai.lum.odinson.extra
+package ai.lum.odinson.extra.utils
 
 import java.util.UUID
 
+import ai.lum.common.ConfigFactory
+import ai.lum.common.ConfigUtils._
+import ai.lum.odinson.{ Document => OdinsonDocument, Sentence => OdinsonSentence, _ }
+import org.clulab.dynet.Utils.initializeDyNet
+import org.clulab.processors.clu.CluProcessor
+import org.clulab.processors.fastnlp.FastNLPProcessor
 import org.clulab.processors.{
   Processor,
   Document => ProcessorsDocument,
   Sentence => ProcessorsSentence
 }
-import ai.lum.odinson.{ Document => OdinsonDocument, Sentence => OdinsonSentence, _ }
-import ai.lum.common.ConfigFactory
-import ai.lum.common.ConfigUtils._
-import edu.cmu.dynet.Initialize
-import org.clulab.dynet.DyNetSync
-import org.clulab.processors.clu.CluProcessor
-import org.clulab.processors.fastnlp.FastNLPProcessor
+import org.clulab.struct.GraphMap
 import org.slf4j.{ Logger, LoggerFactory }
-
-import scala.collection.mutable
 
 object ProcessorsUtils {
 
@@ -74,37 +72,22 @@ object ProcessorsUtils {
     val maybeLemma = s.lemmas.map(lemmas => TokensField(lemmaTokenField, lemmas))
     val maybeEntity = s.entities.map(entities => TokensField(entityTokenField, entities))
     val maybeChunk = s.chunks.map(chunks => TokensField(chunkTokenField, chunks))
-    val maybeDeps = s.dependencies.map(g => GraphField(dependenciesField, g.allEdges, g.roots))
+    // graph that merges ENHANCED_SEMANTIC_ROLES and UNIVERSAL_ENHANCED, if available
+    val maybeDeps = {
+      val graphs = s.graphs match {
+        case hybridCollapsed if hybridCollapsed.contains(GraphMap.HYBRID_DEPENDENCIES) =>
+          hybridCollapsed.get(GraphMap.HYBRID_DEPENDENCIES)
+        case collapsed if collapsed.contains(GraphMap.UNIVERSAL_ENHANCED) =>
+          collapsed.get(GraphMap.UNIVERSAL_ENHANCED)
+        case basic if basic.contains(GraphMap.UNIVERSAL_BASIC) =>
+          basic.get(GraphMap.UNIVERSAL_BASIC)
+        case _ => None
+      }
+      graphs.map(g => GraphField(dependenciesField, g.allEdges, g.roots))
+    }
     val fields =
       Some(raw) :: Some(word) :: List(maybeTag, maybeLemma, maybeEntity, maybeChunk, maybeDeps)
     OdinsonSentence(s.size, fields.flatten)
-  }
-
-  // CluLab processors now uses dynet models, which need to be initialized at first
-  // loading.  These variables and initialization method are for that process.
-  val RANDOM_SEED = 2522620396L // used for both DyNet, and the JVM seed for shuffling data
-  val WEIGHT_DECAY = 1e-5f
-
-  private var IS_DYNET_INITIALIZED = false
-
-  def initializeDyNet(autoBatch: Boolean = false, mem: String = ""): Unit = {
-    DyNetSync.synchronized {
-      if (!IS_DYNET_INITIALIZED) {
-        logger.debug("Initializing DyNet...")
-
-        val params = new mutable.HashMap[String, Any]()
-        params += "random-seed" -> RANDOM_SEED
-        params += "weight-decay" -> WEIGHT_DECAY
-        if (autoBatch) {
-          params += "autobatch" -> 1
-          params += "dynet-mem" -> mem
-        }
-
-        Initialize.initialize(params.toMap)
-        logger.debug("DyNet initialization complete.")
-        IS_DYNET_INITIALIZED = true
-      }
-    }
   }
 
 }